refactor: 테스트 유틸 함수들을 ApplicationTEST.js에서 /src/utils/testUtils.js로 분리
기능별 테스트 코드 작성과 유틸리티 재사용성을 높이기 위해 파일을 분리함
